+++ src/configurations/RGBTConnectConfig.cpp
@@ -4,7 +4,7 @@
 
 #include "RGBTConnectConfig.h"
 
-unsigned long RGBTConnectConfig::MAX_NUM_ITER   = 1e9;
-unsigned long RGBTConnectConfig::MAX_NUM_STATES = 1e9;
-float RGBTConnectConfig::MAX_PLANNING_TIME      = 60;
-int RGBTConnectConfig::NUM_LAYERS               = 5;
+size_t RGBTConnectConfig::MAX_NUM_ITER      = 1e9;
+size_t RGBTConnectConfig::MAX_NUM_STATES    = 1e9;
+float RGBTConnectConfig::MAX_PLANNING_TIME  = 60;
+size_t RGBTConnectConfig::NUM_LAYERS        = 5;
